addressed most of izzy and kristas comments, rest tomorrow morning
