+++ sdk/tables/azure-data-tables/azure/data/tables/_shared_access_signature.py
