+++ sdk/tables/azure-data-tables/azure/data/tables/_table_client.py
@@ -16,6 +16,7 @@ except ImportError:
 from azure.core.paging import ItemPaged
 from azure.core.exceptions import HttpResponseError, ResourceNotFoundError
 from azure.core.tracing.decorator import distributed_trace
+
 from ._deserialize import _convert_to_entity
 from ._entity import TableEntity
 from ._generated import AzureTable
@@ -23,15 +24,10 @@ from ._generated.models import AccessPolicy, SignedIdentifier, TableProperties, 
 from ._serialize import _get_match_headers, _add_entity_properties
 from ._base_client import parse_connection_str
 from ._table_client_base import TableClientBase
-
 from ._serialize import serialize_iso
 from ._deserialize import process_table_error, return_headers_and_deserialized
-
 from ._version import VERSION
-
 from ._models import TableEntityPropertiesPaged, UpdateMode, Table
-
-# from ._shared.response_handlers import return_headers_and_deserialized
 
 
 class TableClient(TableClientBase):
