+++ sdk/tables/azure-data-tables/azure/data/tables/aio/_table_client_async.py
@@ -13,14 +13,14 @@ from azure.core.async_paging import AsyncItemPaged
 from azure.core.exceptions import ResourceNotFoundError, HttpResponseError
 from azure.core.tracing.decorator import distributed_trace
 from azure.core.tracing.decorator_async import distributed_trace_async
+
 from .. import VERSION
 from .._entity import TableEntity
-from .._generated.aio._azure_table_async import AzureTable
+from .._generated.aio import AzureTable
 from .._generated.models import SignedIdentifier, TableProperties, QueryOptions
 from .._models import AccessPolicy, Table
 from .._serialize import serialize_iso
 from .._deserialize import return_headers_and_deserialized, process_table_error
-
 from .._models import UpdateMode
 from .._deserialize import _convert_to_entity
 from .._serialize import _add_entity_properties, _get_match_headers
